Removed some UInt8 casting to resolve two Swift 5 compiler errors.

+++ SwCrypt/SwCrypt.swift
@@ -1345,8 +1345,8 @@ open class CC {
 			var maskedDB = xorData(db, dbMask)
 
 			let zeroBits = 8 * emLength - emBits
-			maskedDB.withUnsafeMutableBytes { maskedDBBytes in
-				maskedDBBytes[0] &= UInt8(0xff >> zeroBits)
+            maskedDB.withUnsafeMutableBytes { maskedDBBytes in
+				maskedDBBytes[0] &= (0xff >> zeroBits)
 			}
 
 			var ret = maskedDB
@@ -1392,7 +1392,7 @@ open class CC {
 			let dbMask = mgf1(digest, seed: mPrimeHash, maskLength: emLength - hash.count - 1)
 			var db = xorData(maskedDB, dbMask)
 			db.withUnsafeMutableBytes { dbBytes in
-				dbBytes[0] &= UInt8(0xff >> zeroBits)
+				dbBytes[0] &= (0xff >> zeroBits)
 			}
 
 			let zeroLength = emLength - hash.count - saltLength - 2
